Low severity - 3 - Audit Report Bug Fix - ZAssetsRegistry - subId is already uint256 Since subId is already declared as uint256, explicit conversion is unnecessary.
Changes are as follows -
- Removed explicit conversion of subId to uint256.

+++ panther-protocol-contracts-ws/contracts/ZAssetsRegistry.sol
@@ -80,9 +80,7 @@ contract ZAssetsRegistry is ImmutableOwnable, IZAssetsRegistry {
         return
             uint160(
                 uint256(
-                    keccak256(
-                        abi.encode(uint256(uint160(token)), uint256(subId))
-                    )
+                    keccak256(abi.encode(uint256(uint160(token)), subId))
                 ) >> 96
             );
     }
